requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
braindecode==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
diffusers==0.34.0
docker-pycreds==0.4.0
docstring-inheritance==2.2.2
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.58.0
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
h5py==3.13.0
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.1.post0
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mne==1.9.0
mpmath==1.3.0
multidict==6.4.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.5
open_clip_torch==2.32.0
packaging==24.2
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.0
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
skorch==1.1.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.6.0
timm==1.0.16
tokenizers==0.21.2
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torcheval==0.0.7
torchinfo==1.8.0
torchmetrics==1.7.3
torchvision==0.22.0+cu128
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
yarl==1.20.1
zipp==3.23.0
